requirements.txt
torch
pillow
opencv-python
numpy
pandas
